requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==2.0.1
pycodestyle==2.6.0
pylint==2.5.3
PyMySQL==0.10.0
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
